requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
comm==0.2.2
configobj==5.0.8
configparser==7.1.0
debugpy==1.8.5
decorator==5.1.1
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.8
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.3.0
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
prov==2.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
pyzmq==26.2.0
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tesseract==0.1.3
tornado==6.4.1
traitlets==5.14.3
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
